// web/src/api/models/TypesDiffLine.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesDiffLine
 */
export interface TypesDiffLine {
    /**
     * 
     * @type {string}
     * @memberof TypesDiffLine
     */
    text: string;
    /**
     * 
     * @type {string}
     * @memberof TypesDiffLine
     */
    type: TypesDiffLineTypeEnum;
}


/**
 * @export
 */
export const TypesDiffLineTypeEnum = {
    Add: 'add',
    Remove: 'remove',
    Same: 'same'
} as const;
export type TypesDiffLineTypeEnum = typeof TypesDiffLineTypeEnum[keyof typeof TypesDiffLineTypeEnum];


/**
 * Check if a given object implements the TypesDiffLine interface.
 */
export function instanceOfTypesDiffLine(value: object): value is TypesDiffLine {
    if (!('text' in value) || value['text'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function TypesDiffLineFromJSON(json: any): TypesDiffLine {
    return TypesDiffLineFromJSONTyped(json, false);
}

export function TypesDiffLineFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesDiffLine {
    if (json == null) {
        return json;
    }
    return {
        
        'text': json['text'],
        'type': json['type'],
    };
}

export function TypesDiffLineToJSON(json: any): TypesDiffLine {
    return TypesDiffLineToJSONTyped(json, false);
}

export function TypesDiffLineToJSONTyped(value?: TypesDiffLine | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'text': value['text'],
        'type': value['type'],
    };
}